requirements.txt
fastapi
pydantic
google-generativeai
python-dotenv
uvicorn
sqlalchemy
pymysql
uvicorn[standard]
mysql-connector-python
